Remove saga from Main container
