+++ packages/strapi-admin/admin/src/config.js
@@ -1,4 +1,5 @@
 // DO NOT MODIFY THESE OPTIONS
 export const LOGIN_LOGO = null;
-export const SHOW_TUTORIALS = true;
+// TODO: uncomment
+export const SHOW_TUTORIALS = false;
 export const SETTINGS_BASE_URL = '/settings';
