+++ packages/strapi-plugin-content-manager/admin/src/containers/Main/utils/createPossibleMainFieldsForModelsAndComponents.js
@@ -1,0 +1,26 @@
+import { get } from 'lodash';
+
+const createPossibleMainFieldsForModelsAndComponents = array => {
+  return array.reduce((acc, current) => {
+    const attributes = get(current, ['schema', 'attributes'], {});
+    const possibleMainFields = Object.keys(attributes).filter(attr => {
+      return ![
+        'boolean',
+        'component',
+        'dynamiczone',
+        'json',
+        'media',
+        'password',
+        'relation',
+        'text',
+        'richtext',
+      ].includes(get(attributes, [attr, 'type'], ''));
+    });
+
+    acc[current.uid] = possibleMainFields;
+
+    return acc;
+  }, {});
+};
+
+export default createPossibleMainFieldsForModelsAndComponents;
